refactor: split _clean_resources_before_push into poly.utils.prepush helpers
The 365-line method mixed eight distinct pre-push adjustments. Each is now a
focused function in poly.utils.prepush (webchat channel enablement, orphaned
variable handling, new-flow grouping, cascade-delete pruning, start-step dummy
workaround, variant defaults/filtering, condition fixes), with project state
passed explicitly. _clean_resources_before_push remains as a thin orchestrator
calling them in the original order, so behaviour and its public contract are
unchanged.
